# docs/swagger/swagger.yaml
basePath: /api/v2
definitions:
  common.ApiVersionData:
    properties:
      app:
        type: string
      commit_hash:
        type: string
      version:
        type: string
    type: object
...
  common.JSONResult:
    properties:
      data:
        description: 数据
      extra:
        description: 数据
      msg:
        description: 消息
        type: string
      success:
        description: 是否成功
        type: boolean
    type: object
...
  entity.CookedComment:
    properties:
      badge_color:
        type: string
      badge_name:
        type: string
      content:
        type: string
      content_marked:
        type: string
      date:
        type: string
      email_encrypted:
        type: string
      id:
        type: integer
      ip_region:
        type: string
      is_allow_reply:
        type: boolean
      is_collapsed:
        type: boolean
      is_pending:
        type: boolean
      is_pinned:
        type: boolean
      link:
        type: string
      nick:
        type: string
      page_key:
        type: string
      page_url:
        type: string
      rid:
        type: integer
      site_name:
        type: string
      ua:
        type: string
      user_id:
        type: integer
      visible:
        type: boolean
      vote_down:
        type: integer
      vote_up:
        type: integer
    type: object
  entity.CookedNotify:
    properties:
      comment_id:
        type: integer
      id:
        type: integer
      is_emailed:
        type: boolean
      is_read:
        type: boolean
      read_link:
        type: string
      user_id:
        type: integer
    type: object
  entity.CookedPage:
    properties:
      admin_only:
        type: boolean
      id:
        type: integer
      key:
        type: string
      pv:
        type: integer
      site_name:
        type: string
      title:
        type: string
      url:
        type: string
      vote_down:
        type: integer
      vote_up:
        type: integer
    type: object
...
  entity.CookedUser:
    properties:
      badge_color:
        type: string
      badge_name:
        type: string
      email:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      link:
        type: string
      name:
        type: string
      receive_email:
        type: boolean
      site_names:
        items:
          type: string
        type: array
      site_names_raw:
        type: string
    type: object
...
  handler.ParamsAdd:
    properties:
      content:
        description: The comment content
        type: string
      email:
        description: The comment email
        type: string
      link:
        description: The comment link
        type: string
      name:
        description: The comment name
        type: string
      page_key:
        description: The comment page_key
        type: string
      page_title:
        description: The comment page_title
        type: string
      rid:
        description: The comment rid
        type: integer
      site_name:
        description: The site name of your content scope
        type: string
      ua:
        description: The comment ua
        type: string
    required:
    - content
    - page_key
    - site_name
    type: object
...
  handler.ParamsLogin:
    properties:
      email:
        description: The user email
        type: string
      name:
        description: The username
        type: string
      password:
        description: The user password
        type: string
    required:
    - email
    - password
    type: object
  handler.ParamsMarkRead:
    properties:
      all_read:
        description: The option if mark all user's notify as read
        type: boolean
      email:
        description: The user email
        type: string
      name:
        description: The username
        type: string
    type: object
...
  handler.ParamsVote:
    properties:
      email:
        description: The user email
        type: string
      name:
        description: The username
        type: string
      site_name:
        description: The site name of your content scope
        type: string
    type: object
  handler.ResponseAdd:
    properties:
      comment:
        $ref: '#/definitions/entity.CookedComment'
    type: object
...
  handler.ResponseGet:
    properties:
      api_version:
        $ref: '#/definitions/common.ApiVersionData'
      comments:
        items:
          $ref: '#/definitions/entity.CookedComment'
        type: array
      page:
        $ref: '#/definitions/entity.CookedPage'
      total:
        type: integer
      total_roots:
        type: integer
      unread:
        items:
          $ref: '#/definitions/entity.CookedNotify'
        type: array
      unread_count:
        type: integer
    type: object
...
  handler.ResponseLogin:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/entity.CookedUser'
    type: object
  handler.ResponseLoginStatus:
    properties:
      is_admin:
        type: boolean
      is_login:
        type: boolean
    type: object
...
  handler.ResponseVote:
    properties:
      down:
        type: integer
      up:
        type: integer
    type: object
info:
  contact:
    email: artalkjs@gmail.com
    name: API Support
    url: https://artalk.js.org
  description: Artalk is a modern comment system based on Golang.
  license:
    name: MIT
    url: https://github.com/ArtalkJS/Artalk/blob/master/LICENSE
  title: Artalk API
  version: "2.0"
paths:
...
  /comments:
    get:
      description: Get a list of comments by some conditions
      parameters:
      - description: The user email
        in: query
        name: email
        type: string
      - description: Enable flat_mode
        in: query
        name: flat_mode
        type: boolean
      - description: The limit for pagination
        in: query
        name: limit
        type: integer
      - description: The username
        in: query
        name: name
        type: string
      - description: The offset for pagination
        in: query
        name: offset
        type: integer
      - description: The comment page_key
        in: query
        name: page_key
        required: true
        type: string
      - description: Search keywords
        in: query
        name: search
        type: string
      - description: The site name of your content scope
        in: query
        name: site_name
        type: string
      - description: Sort by condition
        enum:
        - date_asc
        - date_desc
        - vote
        in: query
        name: sort_by
        type: string
      - description: Message center show type
        enum:
        - all
        - mentions
        - mine
        - pending
        - admin_all
        - admin_pending
        in: query
        name: type
        type: string
      - description: Only show comments by admin
        in: query
        name: view_only_admin
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/handler.ResponseGet'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Comment List
      tags:
      - Comment
    post:
      consumes:
      - application/json
      description: Create a new comment
      parameters:
      - description: The comment data
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/handler.ParamsAdd'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/handler.ResponseAdd'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create Comment
      tags:
      - Comment
...
  /notifies/{comment_id}/{notify_key}/read:
    post:
      consumes:
      - application/json
      description: Mark specific notification as read for user
      parameters:
      - description: The comment id of the notify you want to mark as read
        in: path
        name: comment_id
        required: true
        type: integer
      - description: The key of the notify
        in: path
        name: notify_key
        required: true
        type: string
      - description: The options
        in: body
        name: options
        required: true
        schema:
          $ref: '#/definitions/handler.ParamsMarkRead'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: Mark Notify as Read
      tags:
      - Notify
...
  /user/access_token:
    post:
      consumes:
      - application/json
      description: Login user by name or email
      parameters:
      - description: The user login data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.ParamsLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/handler.ResponseLogin'
              type: object
        "400":
          description: Multiple users with the same email address are matched
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  properties:
                    need_name_select:
                      items:
                        type: string
                      type: array
                  type: object
              type: object
      summary: Get Access Token
      tags:
      - Account
...
  /user/status:
    get:
      description: Get user login status by header Authorization
      parameters:
      - description: The user email
        in: query
        name: email
        type: string
      - description: The username
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/handler.ResponseLoginStatus'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Login Status
      tags:
      - Account
...
  /votes/{type}/{target_id}:
    post:
      consumes:
      - application/json
      description: Vote for a specific comment or page
      parameters:
      - description: The type of vote target
        enum:
        - comment_up
        - comment_down
        - page_up
        - page_down
        in: path
        name: type
        required: true
        type: string
      - description: Target comment or page ID you want to vote for
        in: path
        name: target_id
        required: true
        type: integer
      - description: The vote data
        in: body
        name: vote
        required: true
        schema:
          $ref: '#/definitions/handler.ParamsVote'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/handler.ResponseVote'
              type: object
      summary: Vote
      tags:
      - Vote
...
securityDefinitions:
  ApiKeyAuth:
    description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"